Add chacha-sys crate
