+++ core/Cargo.toml
@@ -49,6 +49,7 @@ serde_derive = "1.0.92"
 serde_json = "1.0.39"
 solana-budget-api = { path = "../programs/budget_api", version = "0.16.0" }
 solana-budget-program = { path = "../programs/budget_program", version = "0.16.0" }
+solana-chacha-sys = { path = "../chacha-sys", version = "0.16.0" }
 solana-client = { path = "../client", version = "0.16.0" }
 solana-config-program = { path = "../programs/config_program", version = "0.16.0" }
 solana-drone = { path = "../drone", version = "0.16.0" }
